Tele-Op with arm
altered the basic tele-op to add simple arm functionality, untested so be very careful. Both arm motors act simultaneously and have the same values, if this is not how the arm is supposed to work, do NOT forget to change! Claw also does not work, change this once details are learned.